requirements.txt
websocket-client
eventlet
flask
flask-socketio
opencv-python
numpy
torch
requests
python-socketio
pyyaml